pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-audio"
dynamic = ["version"]
description = 'Generating audio from research papers'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "Chenghao Mou", email = "mouchenghao@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "redis",
  "pygame",
  "aiohttp",
  "pymupdf",
  "wtpsplit",
  "torch",
  "gradio",
  "loguru",
  "python-dotenv",
  "diskcache",
]

[project.urls]
Documentation = "https://github.com/Chenghao Mou/paper-audio#readme"
Issues = "https://github.com/Chenghao Mou/paper-audio/issues"
Source = "https://github.com/Chenghao Mou/paper-audio"

[tool.hatch.version]
path = "src/paper_audio/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/paper_audio tests}"

[tool.coverage.run]
source_pkgs = ["paper_audio", "tests"]
branch = true
parallel = true
omit = ["src/paper_audio/__about__.py"]

[tool.coverage.paths]
paper_audio = ["src/paper_audio", "*/paper-audio/src/paper_audio"]
tests = ["tests", "*/paper-audio/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
